Working ajax, change to css
Joshua got the ajax working and I stylized it a bit and added a button that when pressed will go to a webform page, that will eventually post and append that new info to the table... to do this we have to figure out how to append to the json file and create a new person object. I will look up how to do this.