More CSS/HTML changes
Changed the table a little bit, stylized the web form, the button on the top right now goes to the web form.

Next is adding to the web form, and trying to save information that is added and then after that is adding that saved and submitted data to the table.

Might also stylize the main page more so that it isnt just a table. I'm feeling pretty good right now though.