validate form function and getting rid of webform
Got rid of the webform because I came to the realization that I think we need PHP to use that, so i got rid of it and just made input boxes and a button that gathers information that is written in the inputs.

Also made  it so that after you hit the button, the values that were given are made into a JSON object.

Now I need to figure out how to put that JSON object into that table and hopefully we are done then.