Got it to post to the table
Got the form to post to the table.

Now when we first load into the page it generates an error, however if we add something to the page, it wont generage the error..

another problem is that we can only ever add 1 item to the table but i mean, i dont think we can do anything about it since we arent using php, or a database.