Final update(I think)
Cant fix the error when you first come to the website, have a way to have it not show but its just by submitting a webform first and then going to the homepage. (I can make it look like it doesnt happen while presenting)

Other than that It is pretty much done, I commented everything, I can do the explaining tonight.

But I'm just done with the project for now, don't think we have the knowledge and tools to do what the orginal plan was.